+++ src/app/app.component.spec.ts
@@ -1,10 +1,43 @@
 import { TestBed, async } from '@angular/core/testing';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { MatToolbarModule, MatSelectModule, MatSidenavModule, MatIconModule, MatButtonModule, MatMenuModule } from '@angular/material';
+import { TranslateModule } from '@ngx-translate/core';
+import { PipThemesModule } from 'pip-webui2-themes';
+
 import { AppComponent } from './app.component';
+import { AppRoutingModule } from './app-routing.module';
+import { ExamplesListModule } from './examples-list/examples-list.module';
+import { CollageExampleModule } from './collage-example/collage-example.module';
+import { PictureEditExampleModule } from './picture-edit-example/picture-edit-example.module';
+import { PictureExampleModule } from './picture-example/picture-example.module';
+import { PictureListEditExampleModule } from './picture-list-edit-example/picture-list-edit-example.module';
+
 describe('AppComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [
         AppComponent
+      ],
+      imports: [
+        NoopAnimationsModule,
+        FlexLayoutModule,
+        MatToolbarModule,
+        MatSelectModule,
+        MatSidenavModule,
+        MatIconModule,
+        MatButtonModule,
+        MatMenuModule,
+        TranslateModule.forRoot(),
+
+        PipThemesModule,
+
+        AppRoutingModule,
+        ExamplesListModule,
+        CollageExampleModule,
+        PictureEditExampleModule,
+        PictureExampleModule,
+        PictureListEditExampleModule
       ],
     }).compileComponents();
   }));
@@ -12,16 +45,5 @@ describe('AppComponent', () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.debugElement.componentInstance;
     expect(app).toBeTruthy();
-  }));
-  it(`should have as title 'pip-webui2-dates-app'`, async(() => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.debugElement.componentInstance;
-    expect(app.title).toEqual('pip-webui2-dates-app');
-  }));
-  it('should render title in a h1 tag', async(() => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector('h1').textContent).toContain('Welcome to pip-webui2-dates-app!');
   }));
 });