+++ src/app/collage-example/collage-example.component.spec.ts
@@ -1,4 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+import { TranslateModule } from '@ngx-translate/core';
+import { PipCollageModule } from 'pip-webui2-pictures';
 
 import { CollageExampleComponent } from './collage-example.component';
 
@@ -8,9 +11,15 @@ describe('CollageExampleComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ CollageExampleComponent ]
+      declarations: [CollageExampleComponent],
+      imports: [
+        CommonModule,
+        TranslateModule.forRoot(),
+
+        PipCollageModule
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {