+++ src/app/examples-list/examples-list.component.spec.ts
@@ -1,4 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { FormsModule } from '@angular/forms';
+import { MatListModule } from '@angular/material';
+import { RouterModule } from '@angular/router';
 
 import { ExamplesListComponent } from './examples-list.component';
 
@@ -8,9 +13,16 @@ describe('ExamplesListComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ ExamplesListComponent ]
+      declarations: [ExamplesListComponent],
+      imports: [
+        CommonModule,
+        FlexLayoutModule,
+        FormsModule,
+        MatListModule,
+        RouterModule
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {