+++ src/app/picture-list-edit-example/picture-list-edit-example.component.spec.ts
@@ -1,4 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { CommonModule } from '@angular/common';
+import { TranslateModule } from '@ngx-translate/core';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { PipPictureListEditModule } from 'pip-webui2-pictures';
 
 import { PictureListEditExampleComponent } from './picture-list-edit-example.component';
 
@@ -8,9 +12,16 @@ describe('PictureListEditExampleComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ PictureListEditExampleComponent ]
+      declarations: [PictureListEditExampleComponent],
+      imports: [
+        NoopAnimationsModule,
+        CommonModule,
+        TranslateModule.forRoot(),
+
+        PipPictureListEditModule
+      ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {